rewrite machine processor handling
This commit rewrites the logic for applying processor logic to machines. Instead of replacing the in-world machine instances, the mod now uses Harmony to intercept machine methods and invoke singleton processors with the machine instance. This is much less error-prone, fixes save errors when machines are in some custom locations due to race conditions, fixes other mod compatibility issues, and fixes a machine duplication bug.